Player progress manager
Created a new player progress manager class to consolidate functionality about player data. The new class maintains a server-side master player progress object, while maintaining individual files for each user to track progress in the scavenger hunt. This new master progress object can be used in a future update to generate a leaderboard for the current session, tracking which players have turned in the most items in the hunt.