Leaderboard Implemented
Aside from the leaderboard itself, which is displayed at the bottom of the info panel in the quest hub block, there is also a new command "/leaderboard" which will show you the leaderboard status in the chat  area. I've tried to stop using UUID wherever possible and just use player names directly, which led to several other changes in code, but those are the main features.