user defined checklist size
when the /starthunt command is used, the user can now pass an optional integer argument defining how many items are in the hunt. If no argument is given, the hunt will default to 10 items. If the number exceeds the amount of available items from the item pool, a chat message is given which informs the user of the maximum number of recipes.