packet management
Implemented a basic packet manager, so information can be sent between the clients and the server when necessary now. This helps correct features that work in single player, but are broken in multiplayer and either cause massive errors or just crash the client or server outright. The first error corrected using the packet manager was making sure that 'isHuntStarted' has its state on the server sent to all connected clients whenever its updated, or when a client logs in to the server. That fixed an error where in multiplayer, the player was unable to interact with any part of the world even after the hunt was started. But now a null pointer exception is thrown, crashing the server if the user attempts to use the quest hub block, because the player progress isn't yet being sent to the clients for the purpose of rendering on the info panel. This will be corrected soon.