syncing player progress
Now, the master player progress object which holds progress for all players in the server is properly syncing from the client to the server. As a result most features of the mod are working in multiplayer, however the leaderboard is still broken. It's probably being created on the server and needs to be synced with clients as well.