synced leaderboard
The leaderboard is now working in multiplayer, but it seems that the player progress isnt actually being tracked properly. The leaderboard will change to reflect updates to the number of turn ins, but the completed and incomplete items in the leaderboard are not updating.

On a side note, it does seem that time is correctly syncing, and along with that, the ability to see server spawned entities now that the ishuntstarted variable is syncing properly.