custom spawn protection
Spawn protection was causing items to be consumed from the player's inventory when they were dropped in the spawn protection radius, and cancelling block placement events was desyncing the client and server copies of the player inventories. So I implemented custom spawn protection. It needs to be cleaned up a bit, but now at least if a player tries to drop his fresh quest hub block in the spawn protection area, it doesnt just appear to disappear while other players still see them holding it, like it used to.

Now there seems to be an issue with the player progress data not propagating when items are turned in, but then being correct on player relogging in multiplayer, so thats next up.