delay return block
the method i was using to ensure that players who place blocks in the spawn protection area got their blocks back wasn't working consistently during multiplayer, and the code i was using to try and return the block needed some cleaning up anyway.

If it takes the server more than a half second to remove the block from the player's inventory after they place it, then they arent going to get the block back.

This isnt ideal, but at least in vanilla in my setup, it seems to be sufficient to prevent the player from losing their quest hub block right after they get it by trying to place it in the spawn zone