progress packets and spawn protection
The main goal for this update was to correct progress syncing between players and the server when an item was turned in. That's been handled and wasn't too difficult.

Along the way i discovered that the action cancellation for the custom spawn protection was very broken still, but it looks like i finally corrected that by just setting the event to cancelled and by reacting to onPlayerInteract to detect block placement rather than onBlockPlace.

onBlockPlace won't even fire if the server default spawn protection cancels the block placement first, but onPlayerInteract can be cancelled on the client, before the block placement goes through, which automatically preserves the player's item, so now no packets need to be sent to sync the inventory state, blocks dont need to be retroactively removed from the world, and we don't need to delay adding the block back to the player's inventory. All of that was replaced by simple cancelling the event on the client in the onPlayerInteract event.

Next will probably be syncing the hunt winner with the clients and server when appropriate.