Hunt Winner Sync
Now, the hunt winner should sync in multiplayer so that the leaderboard command and the related section of the info panel in the quest block should let the player know who the winner is when someone has won.

Additionally, the messages regarding the hunt starting and ending will now broadcast to all connected players in multiplayer and not just the person issuing the starthunt command, or just the person who wins the hunt.

The fireworks fanfare was already working in multiplayer and happens for all clients. They can cause damage and are semi randomly placed, but I might leave it that way unless people complain. (I think its fine)